LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-4.7.2.1-install-event-section-expected.txt


FAIL InstallEvent interface: existence and properties of interface object assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent interface object length assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent interface object name assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent interface: attribute activeWorker assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent interface: operation replace() assert_own_property: self does not have own property "InstallEvent" expected property "InstallEvent" missing
FAIL InstallEvent must be primary interface of throw new Error ('No object defined for the InstallEvent interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the InstallEvent interface"
FAIL Stringification of throw new Error ('No object defined for the InstallEvent interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the InstallEvent interface"
FAIL InstallEvent interface: throw new Error ('No object defined for the InstallEvent interface') must inherit property "activeWorker" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the InstallEvent interface"
FAIL InstallEvent interface: throw new Error ('No object defined for the InstallEvent interface') must inherit property "replace()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the InstallEvent interface"
